+++ sources/Sprite.hpp
@@ -123,8 +123,8 @@
 		};
 		indexCount = sizeof(index) / sizeof(index[0]);
 
-		int x = App::GetWindowSize().x;
-		int y = App::GetWindowSize().y;
+		int x = static_cast<int>(App::GetWindowSize().x);
+		int y = static_cast<int>(App::GetWindowSize().y);
 
 		D3D11_BUFFER_DESC vertexBufferDesc = {};
 		vertexBufferDesc.Usage = D3D11_USAGE_DEFAULT;
